fix: resolved main issues in PR #5
Still need to be tested
